+++ src/biml.ts
@@ -134,13 +134,13 @@ export class BiML extends BiDoc {
               (options?.def?.showBrackets ? "]]" : "")
             }</span>`,
           // ref renderer
-          (ref, def) => {
-            const span = `<span id="${this.refIdGenerator(ref, def)}">${
+          (ref) => {
+            const span = `<span id="${this.refIdGenerator(ref)}">${
               (options?.ref?.showBrackets ? "[[" : "") +
               ref.name + // don't use def.name here, because it may be an alias
               (options?.ref?.showBrackets ? "]]" : "")
             }</span>`;
-            return `<a href="${def.path}#${def.id}">${span}</a>`;
+            return `<a href="${ref.def.path}#${ref.def.id}">${span}</a>`;
           }
         ),
       };